SE-1434 add reconnect arg to connect() and new attachPrincipalPolicy API
* add a default reconnect param to the connect api
* new attachPrinicipalPolicy API which can be used to programatically add a iot policy to a cognito user
