+++ lib/flutter_amazon_freertos_plugin.dart
@@ -8,5 +8,7 @@ import "package:logging/logging.dart";
 import "package:plugin_scaffold/plugin_scaffold.dart";
 
 part "src/flutter_amazon_freertos_plugin.dart";
-part "src/models.dart";
 part "src/utils.dart";
+
+part 'src/models/bluetooth.model.dart';
+part 'src/models/aws.model.dart';
