+++ lib/src/flutter_amazon_freertos_plugin.dart
@@ -1,63 +1,70 @@
 part of flutter_amazon_freertos_plugin;
 
 typedef OnBluetoothStateChange(BluetoothState bluetoothState);
+
 class FlutterAmazonFreeRTOSPlugin {
-static final logger = Logger("FlutterAmazonFreeRTOSPlugin");
-static const pkgName = "nl.qwic.plugins.flutter_amazon_freertos_plugin";
-final MethodChannel channel = const MethodChannel(pkgName);
-static FlutterAmazonFreeRTOSPlugin _instance = new FlutterAmazonFreeRTOSPlugin();
-static FlutterAmazonFreeRTOSPlugin get instance => _instance;
+    static final logger = Logger("FlutterAmazonFreeRTOSPlugin");
+    static const pkgName = "nl.qwic.plugins.flutter_amazon_freertos_plugin";
+    final MethodChannel channel = const MethodChannel(pkgName);
+    static FlutterAmazonFreeRTOSPlugin _instance = new FlutterAmazonFreeRTOSPlugin();
+    static FlutterAmazonFreeRTOSPlugin get instance => _instance;
 
-Future<BluetoothState> get bluetoothState async {
-    final int bluetoothState = await channel.invokeMethod("bluetoothState");
-    return BluetoothState.values[bluetoothState];
-}
-
-void registerBluetoothStateChangeCallback(OnBluetoothStateChange onBluetoothStateChange) {
-    const _method = "bluetoothStateChangeCallback";
-
-    if (onBluetoothStateChange == null) {
-        PluginScaffold.removeCallHandlersWithName(channel, _method);
-        return;
+    Future<BluetoothState> get bluetoothState async {
+        final int bluetoothState = await channel.invokeMethod("bluetoothState");
+        return BluetoothState.values[bluetoothState];
     }
 
-    PluginScaffold.setCallHandler(channel, _method, (bluetoothState) {
-        onBluetoothStateChange(BluetoothState.values[bluetoothState]);
-    });
-}
+    void registerBluetoothStateChangeCallback(OnBluetoothStateChange onBluetoothStateChange) {
+        const _method = "bluetoothStateChangeCallback";
 
-Stream<FreeRTOSDevice> startScanForDevices({ List<String> serviceUUIDS: const [], int scanDuration = 0}) {
-    return PluginScaffold.createStream(channel, "startScanForDevices", {
-        "serviceUUIDS": serviceUUIDS,
-        "scanDuration": scanDuration,
-    }).map((value) {
-        return FreeRTOSDevice.fromJson(value);
-    }).handleError((e) {
-        Exception(e);
-    });
-}
+        if (onBluetoothStateChange == null) {
+            PluginScaffold.removeCallHandlersWithName(channel, _method);
+            return;
+        }
 
-Future<void> stopScanForDevices() async {
-    await channel.invokeMethod("stopScanForDevices");
-}
+        PluginScaffold.setCallHandler(channel, _method, (bluetoothState) {
+            onBluetoothStateChange(BluetoothState.values[bluetoothState]);
+        });
+    }
 
-Stream<FreeRTOSDevice> rescanForDevices({List<String> serviceUUIDS = const [], int scanDuration = 0}) {
-    return PluginScaffold.createStream(channel, "rescanForDevices", {
-        "servieUUIDS": serviceUUIDS,
-        "scanDuration": scanDuration,
-    }).map((value) {
-        return FreeRTOSDevice.fromJson(value);
-    }).handleError((e) {
-        Exception(e);
-    });
-}
+    Stream<FreeRTOSDevice> startScanForDevices({ List<String> serviceUUIDS: const [], int scanDuration = 0}) {
+        return PluginScaffold.createStream(channel, "startScanForDevices", {
+            "serviceUUIDS": serviceUUIDS,
+            "scanDuration": scanDuration,
+        }).map((value) {
+            return FreeRTOSDevice.fromJson(value);
+        }).handleError((e) {
+            Exception(e);
+        });
+    }
 
-Future<List<FreeRTOSDevice>> get discoveredDevices async {
-    final List devices = await channel.invokeListMethod("listDiscoveredDevices");
-    return List<FreeRTOSDevice>.from(devices.map((device) {
-        return FreeRTOSDevice.fromJson(device);
-    }));
-}
+    Future<void> stopScanForDevices() async {
+        await channel.invokeMethod("stopScanForDevices");
+    }
 
-//String get pkgName => pkgName;
+    Stream<FreeRTOSDevice> rescanForDevices({List<String> serviceUUIDS = const [], int scanDuration = 0}) {
+        return PluginScaffold.createStream(channel, "rescanForDevices", {
+            "servieUUIDS": serviceUUIDS,
+            "scanDuration": scanDuration,
+        }).map((value) {
+            return FreeRTOSDevice.fromJson(value);
+        }).handleError((e) {
+            Exception(e);
+        });
+    }
+
+    Future<List<FreeRTOSDevice>> get discoveredDevices async {
+        final List devices = await channel.invokeListMethod("listDiscoveredDevices");
+        return List<FreeRTOSDevice>.from(devices.map((device) {
+            return FreeRTOSDevice.fromJson(device);
+        }));
+    }
+
+
+    // IOS Only
+    // Attaches IOT policy to a cognito user
+    Future<void> attachPrincipalPolicy({ String policyName, AWSRegionType awsRegion }) async {
+        await channel.invokeMethod("attachPrinicipalPolicy", { "policyName": policyName, awsRegion: awsRegion });
+    }
+
 }
