+++ lib/src/models/aws.model.dart
@@ -1,0 +1,29 @@
+part of flutter_amazon_freertos_plugin;
+
+// Enums for AWS regions.
+// For information about which regions are supported for each service, see the linked website:
+// http://docs.aws.amazon.com/general/latest/gr/rande.html
+enum AWSRegionType {
+    UNKNOWN,
+    USEAST1,
+    USEAST2,
+    USWEST1,
+    USWEST2,
+    EUWEST1,
+    EUWEST2,
+    EUCENTRAL1,
+    APSOUTHEAST1,
+    APNORTHEAST1,
+    APNORTHEAST2,
+    APSOUTH1,
+    SAEAST1,
+    CNNORTH1,
+    CACENTRAL1,
+    USGOVWEST1,
+    CNNORTHWEST1,
+    EUWEST3,
+    USGOVEAST1,
+    EUNORTH1,
+    APEAST1,
+    MESOUTH1
+}
